+++ src/biz/bokhorst/xprivacy/XMediaRecorder.java
@@ -9,7 +9,7 @@ public class XMediaRecorder extends XHook {
 	private Methods mMethod;
 
 	private XMediaRecorder(Methods method, String restrictionName) {
-		super(restrictionName, method.name(), null);
+		super(restrictionName, method.name(), "MediaRecorder." + method.name());
 		mMethod = method;
 	}
 
@@ -17,32 +17,25 @@ public class XMediaRecorder extends XHook {
 		return "android.media.MediaRecorder";
 	}
 
-	// public void setOutputFile(FileDescriptor fd)
-	// public void setOutputFile(String path)
+	// public native void start()
 	// frameworks/base/media/java/android/media/MediaRecorder.java
 	// http://developer.android.com/reference/android/media/MediaRecorder.html
 
 	private enum Methods {
-		setOutputFile
+		start
 	};
 
 	public static List<XHook> getInstances() {
 		List<XHook> listHook = new ArrayList<XHook>();
-		listHook.add(new XMediaRecorder(Methods.setOutputFile, PrivacyManager.cMedia));
+		listHook.add(new XMediaRecorder(Methods.start, PrivacyManager.cMedia));
 		return listHook;
 	}
 
 	@Override
 	protected void before(XParam param) throws Throwable {
-		if (mMethod == Methods.setOutputFile) {
-			if (param.args.length > 0)
-				if (param.args[0] instanceof String) {
-					if (isRestrictedExtra(param, (String) param.args[0]))
-						param.setResult(null);
-				} else {
-					if (isRestricted(param))
-						param.setResult(null);
-				}
+		if (mMethod == Methods.start) {
+			if (isRestricted(param))
+				param.setResult(null);
 
 		} else
 			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());